ORG: Shuffling around ReallyEqualMixin
