+++ skbio/alignment/tests/test_tabular_msa.py
@@ -23,7 +23,8 @@ import scipy.stats
 from skbio import Sequence, DNA, RNA, Protein, TabularMSA
 from skbio.sequence import GrammaredSequence
 from skbio.util._decorator import classproperty, overrides
-from skbio.metadata._testing import (ReallyEqualMixin, MetadataMixinTests,
+from skbio.util._testing import ReallyEqualMixin
+from skbio.metadata._testing import (MetadataMixinTests,
                                      PositionalMetadataMixinTests)
 from skbio.util._testing import (assert_index_equal,
                                  assert_data_frame_almost_equal)
