+++ skbio/metadata/tests/test_mixin.py
@@ -12,7 +12,8 @@ import unittest
 
 from skbio.metadata._mixin import (MetadataMixin, PositionalMetadataMixin)
 from skbio.util._decorator import overrides
-from skbio.metadata._testing import (ReallyEqualMixin, MetadataMixinTests,
+from skbio.util._testing import ReallyEqualMixin
+from skbio.metadata._testing import (MetadataMixinTests,
                                      PositionalMetadataMixinTests)
